electionsproject/requirements.txt
alabaster==0.7.7
args==0.1.0
Babel==2.2.0
CacheControl==0.11.7
cement==2.10.2
cffi==1.5.2
clint==0.5.1
cryptography==1.3.1
defusedxml==0.4.1
Django==1.9
django-debug-toolbar==1.4
django-nested-admin==3.0.12
django-rest-swagger==0.3.4
django-tastypie==0.13.1
django-tastypie-swagger==0.1.4
djangorestframework==3.3.2
docutils==0.12
elex==2.4.1
enum34==1.1.2
futures==3.0.5
idna==2.1
ipaddress==1.0.16
Jinja2==2.8
lockfile==0.12.2
lxml==3.7.2
Markdown==2.6.5
MarkupSafe==0.23
ndg-httpsclient==0.4.0
nose2==0.5.0
pbr==1.8.1
pep8==1.7.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
Pygments==2.1
pymongo==3.3.0
pyOpenSSL==16.0.0
python-dateutil==2.2
python-mimeparse==0.1.4
python-monkey-business==1.0.0
pytz==2015.7
PyYAML==3.11
requests==2.11.1
schedule==0.3.2
six==1.10.0
slacker==0.9.9
snowballstemmer==1.2.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
sqlparse==0.1.18
stevedore==1.10.0
tinys3==0.1.12
ujson==1.35
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
